- Refactored all instances of "unsigned int" to uint64_t - Changed cuda_init for initializers to use bool instead of int - Added initial support for native myriad allocator - Changed several usages of naked pointers to compile-time arrays - Added hard-coded function to calculate program heap size at runtime

+++ HHKCurrMechanism.cu
@@ -15,7 +15,6 @@
 
 extern "C"
 {
-    #include "myriad_debug.h"
 	#include "MyriadObject.h"
     #include "Compartment.h"
 	#include "HHSomaCompartment.h"
@@ -29,14 +28,12 @@ extern "C"
 __device__ __constant__ struct HHKCurrMechanism* HHKCurrMechanism_dev_t;
 __device__ __constant__ struct HHKCurrMechanismClass* HHKCurrMechanismClass_dev_t;
 
-__device__ double HHKCurrMechanism_cuda_mech_fun(
-    void* _self,
-	void* pre_comp,
-	void* post_comp,
-	const double dt,
-	const double global_time,
-	const unsigned int curr_step
-	)
+__device__ double HHKCurrMechanism_cuda_mech_fun(void* _self,
+                                                 void* pre_comp,
+                                                 void* post_comp,
+                                                 const double dt,
+                                                 const double global_time,
+                                                 const uint64_t curr_step)
 {
 	struct HHKCurrMechanism* self = (struct HHKCurrMechanism*) _self;
 	const struct HHSomaCompartment* c1 = (const struct HHSomaCompartment*) pre_comp;
